requirements.txt
astroid==3.0.1
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
branca==0.7.0
bs4==0.0.1
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
folium==0.15.0
greenlet==3.0.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
numpy==1.26.1
opencage==2.3.0
package-name==0.1
packaging==23.2
pandas==2.1.3
phonenumbers==8.13.24
pip-tools==7.3.0
platformdirs==3.11.0
protobuf==4.21.12
pylint==3.0.2
PyMySQL==1.1.0
pyodbc==5.0.1
pyproject-toml==0.0.10
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
setuptools==69.0.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
toml==0.10.2
tomlkit==0.12.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
Werkzeug==3.0.1
wheel==0.41.3
